Fix missing settings on migration
Some defaults of the WebTop application are stored in the PostgreSQL
database. To create new defaults or modify existing ones, the settings
are changed in the initialization script during a fresh installation or
via the `update-module.d` step during an update.

However, during migration, the step in `update-module.d` is not
executed, and the database initialization script is not invoked because
the database is restored from a dump of the old one. As a result, the
new default settings are not adjusted.

Since the `configure-module` action is launched as the final step of
migration, as a workaround, the setting adjustment step from
`update-module.d` is also included in `configure-module`.